+++ packages/instantsearch.js/src/lib/utils/isIndexWidget.ts
@@ -1,7 +1,9 @@
+import { indexWidgetTypes } from '../../types';
+
 import type { Widget, IndexWidget } from '../../types';
 
 export function isIndexWidget(
   widget: Widget | IndexWidget
 ): widget is IndexWidget {
-  return widget.$$type === 'ais.index';
+  return indexWidgetTypes.includes(widget.$$type as (typeof indexWidgetTypes)[number]);
 }