+++ packages/vue-instantsearch/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.26.3](https://github.com/algolia/instantsearch/compare/vue-instantsearch@4.26.2...vue-instantsearch@4.26.3) (2026-05-12)
+
+**Note:** Version bump only for package vue-instantsearch
+
+
+
+
+
+## [4.26.1](https://github.com/algolia/instantsearch/compare/vue-instantsearch@4.26.0...vue-instantsearch@4.26.1) (2026-05-05)
+
+
+### Bug Fixes
+
+* **feeds:** rename searchScope to isolated ([#7013](https://github.com/algolia/instantsearch/issues/7013)) ([8c43f3a](https://github.com/algolia/instantsearch/commit/8c43f3a4741ebbf52a331fa3f65006d4d0695df7))
+
+
+
+
+
+# [4.26.0](https://github.com/algolia/instantsearch/compare/vue-instantsearch@4.25.1...vue-instantsearch@4.26.0) (2026-05-05)
+
+
+### Features
+
+* **composition:** add Feeds widget to support multifeed ([#7004](https://github.com/algolia/instantsearch/issues/7004)) ([ecc40d9](https://github.com/algolia/instantsearch/commit/ecc40d98dd19922b510dd3d769fc490519cede64)), closes [#6954](https://github.com/algolia/instantsearch/issues/6954)
+
+
+
+
+
 ## [4.25.1](https://github.com/algolia/instantsearch/compare/vue-instantsearch@4.25.0...vue-instantsearch@4.25.1) (2026-04-28)
 
 **Note:** Version bump only for package vue-instantsearch